starting to similfy web accessibility code
